+++ packages/uno-serverless-aws/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./events";
 export * from "./handlers";
 export * from "./middlewares";
 export * from "./services";